Difficulty Window Focus on Shell/Run Window
Window with image runs, RFID card will not scan and verify against text file in line 64. Have to leave open string on line 62 to allow for some kind of functionality. If line 62 is rfid_presented = input(), then the only way this works if I click inside of shell and scan card manually. I am a bit confused how to fix this...